feat: Add home view and related navigation structure for movie categories and favorites
