+++ lib/config/app_router.dart
@@ -1,11 +1,16 @@
 import 'package:cinemapedia_app/presentation/screens/screens.dart';
 import 'package:go_router/go_router.dart';
 
-final appRouter = GoRouter(initialLocation: '/', routes: [
+final appRouter = GoRouter(initialLocation: '/home/0', routes: [
   GoRoute(
-      path: '/',
+      path: '/home/:page',
       name: HomeScreen.name,
-      builder: (context, state) => const HomeScreen(),
+      builder: (context, state) {
+        final int page = int.tryParse(state.pathParameters['page'] ?? '0') ?? 0;
+        return HomeScreen(
+          pageIndex: page,
+        );
+      },
       routes: [
         GoRoute(
           path: 'movie/:id',
@@ -18,4 +23,8 @@ final appRouter = GoRouter(initialLocation: '/', routes: [
           },
         )
       ]),
+  GoRoute(
+    path: '/',
+    redirect: (context, state) => '/home/0',
+  ),
 ]);
