+++ lib/presentation/providers/movies/initial_loading_provider.dart
@@ -4,7 +4,7 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'initial_loading_provider.g.dart';
 
-@riverpod
+@Riverpod(keepAlive: true)
 bool initialLoading(Ref ref) {
   final step1 = ref.watch(nowPlayingMoviesNotifierProvider).isEmpty;
   final step2 = ref.watch(popularMoviesNotifierProvider).isEmpty;
