+++ lib/presentation/views/movies/categories_view.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class CategoriesView extends StatelessWidget {
+  const CategoriesView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Categorías'),
+    );
+  }
+}
