+++ lib/presentation/views/movies/favorites_view.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class FavoritesView extends StatelessWidget {
+  const FavoritesView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('No hay favoritos'),
+    );
+  }
+}
