+++ lib/presentation/views/views.dart
@@ -1,0 +1,3 @@
+export 'movies/categories_view.dart';
+export 'movies/favorites_view.dart';
+export 'movies/home_view.dart';
