+++ lib/presentation/widgets/movies/movie_horizontal_listview.dart
@@ -231,7 +231,7 @@ class _PosterImage extends StatelessWidget {
 
     return GestureDetector(
       onTap: () {
-        context.push('/movie/${movie.id}');
+        context.push('/home/0/movie/${movie.id}');
       },
       child: imageContainer,
     );
