+++ lib/presentation/widgets/movies/movies_slideshow.dart
@@ -53,7 +53,7 @@ class _Slide extends StatelessWidget {
 
     return GestureDetector(
       onTap: () {
-        context.push('/movie/${movie.id}');
+        context.push('/home/0/movie/${movie.id}');
       },
       child: Padding(
         padding: const EdgeInsets.only(bottom: 30),
