+++ lib/presentation/widgets/shared/custom_appbar.dart
@@ -52,7 +52,7 @@ class CustomAppbar extends ConsumerWidget {
                                   initialMovies: searchedMovies))
                           .then((movie) {
                         if (movie == null || !context.mounted) return;
-                        context.push('/movie/${movie.id}');
+                        context.push('/home/0/movie/${movie.id}');
                       });
                     },
                     icon: Icon(
